Transfered Modal to Utils and opimized it for multipile use
