+++ src/__tests__/ProductView/Modal.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Modal from '../../ProductView/TabBar/QueryInterface/Modal/Modal.js';
+import Modal from '../../Utils/Modal/Modal.js';
 import renderer from 'react-test-renderer';
 let instance;
 
@@ -8,9 +8,9 @@ test('On save callback test', () => {
     renderer.create(
         <Modal
             ref={(child) => {instance = child}}
-            onSave={callbackMock}/>
+            onSubmit={callbackMock}/>
     );
-    instance.onSave();
+    instance.onSubmit();
     expect(callbackMock).toBeCalled();
 
 });
